Refactor basetype extensions
