+++ src/Cosmos.Extensions/System/Extensions.Array.cs
@@ -1,8 +1,6 @@
-using System;
 using System.Collections;
 
-// ReSharper disable once CheckNamespace
-namespace Cosmos
+namespace System
 {
     public static partial class BaseTypeExtensions
     {
